+++ install.py
@@ -31,7 +31,6 @@ class BasicInstaller(ExtensionInstaller):
                     "bin/user",
                     [
                         "bin/user/weewx_wdc.py",
-                        "bin/user/weewx_wdc_forecast.py",
                     ],
                 ),
                 (
@@ -121,6 +120,9 @@ class BasicInstaller(ExtensionInstaller):
                         "skins/weewx-wdc/includes/icons/forecast/SC.svg",
                         "skins/weewx-wdc/includes/icons/forecast/rain.svg",
                         "skins/weewx-wdc/includes/icons/forecast/snow.svg",
+                        "skins/weewx-wdc/includes/icons/forecast/sleet.svg",
+                        "skins/weewx-wdc/includes/icons/forecast/fog.svg",
+                        "skins/weewx-wdc/includes/icons/forecast/haze.svg",
                         "skins/weewx-wdc/includes/icons/forecast/snow--scattered.svg",
                         "skins/weewx-wdc/includes/icons/forecast/rain--scattered.svg",
                         "skins/weewx-wdc/includes/icons/forecast/thunderstorm.svg",